Configured the slimming file and modules file
- Using the 'menuselect-tree' file from the Asterisk 16.6.1 tarball, I
created a default "module_slimming_1_16.conf' file. This shows all the
modules that can be configured for Asterisk. We default to not auto load
any modules then we will pick and choose what modules we want standard,
what a user might need and what will not be used from the GUI. The
modules.conf was then overwritten with this 'default' file and I
manually selected modules to get the SBC running as slim as possible.